# crates/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "easyImmerse API",
    "description": "The REST API of easyImmerse.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/conversion-cache": {
      "get": {
        "tags": [
          "conversions"
        ],
        "operationId": "getConversionCache",
        "responses": {
          "200": {
            "description": "The cache's usage and limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionCacheStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "This server does not convert media (code `conversion_unavailable`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/conversion-cache/budget": {
      "put": {
        "tags": [
          "conversions"
        ],
        "summary": "Sets how large the cache may grow, keeps the choice for later runs, and answers with the status under it.",
        "operationId": "setConversionCacheBudget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversionCacheBudget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The cache's usage and limits under the new budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionCacheStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "This server does not convert media (code `conversion_unavailable`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/conversion-cache/clear": {
      "post": {
        "tags": [
          "conversions"
        ],
        "summary": "Removes every cached conversion that is not in use right now.",
        "operationId": "clearConversionCache",
        "responses": {
          "200": {
            "description": "The cache's usage and limits after clearing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionCacheStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "421": {
            "description": "Unexpected Host header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "This server does not convert media (code `conversion_unavailable`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable, machine-readable identifier such as `unauthorized`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ConversionCacheBudget": {
        "type": "object",
        "description": "How large the conversion cache may grow. `null` lets the budget follow the disk's size.",
        "properties": {
          "budget_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ConversionCacheStatus": {
        "type": "object",
        "description": "Sizes are in bytes.",
        "required": [
          "usage_bytes",
          "limit_bytes",
          "budget_bytes",
          "free_bytes",
          "space_low"
        ],
        "properties": {
          "budget_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "The size the cache may grow to when disk space allows.",
            "minimum": 0
          },
          "chosen_budget_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The budget the user chose, or None while the budget follows the disk's size.",
            "minimum": 0
          },
          "free_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "limit_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "The size the cache may currently grow to: the budget, reduced when free space is short.",
            "minimum": 0
          },
          "space_low": {
            "type": "boolean",
            "description": "True when free disk space, not the budget, limits the cache."
          },
          "usage_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}